Remove eval, use evalWithCost in tests
