+++ src/main/kotlin/moirai/eval/EvalWithCost.kt
@@ -1,13 +1,30 @@
 package moirai.eval
 
-import moirai.composition.CompilerFrontend
-import moirai.composition.ExecutionArtifacts
-import moirai.composition.ExecutionCache
-import moirai.composition.pluginMap
+import moirai.composition.*
 import moirai.semantics.core.Architecture
 import moirai.transport.TransportAst
 
 data class EvalWithCostResult(val value: Value, val cost: Long, val alternativeCosts: Map<Architecture, Long>)
+
+fun evalWithCost(
+    source: String,
+    architecture: Architecture,
+    sourceStore: SourceStore
+): EvalWithCostResult {
+    val frontend = CompilerFrontend(architecture, sourceStore)
+
+    val executionArtifacts = frontend.fullCompileWithTopologicalSort(source)
+
+    val globalScope = ValueTable(NullValueTable)
+    val evalVisitor = EvalAstVisitor(architecture, globalScope, mapOf())
+
+    val executionScope = ValueTable(globalScope)
+    val value = executionArtifacts.processedAst.accept(evalVisitor, EvalContext(executionScope, mapOf()))
+    val cost = executionArtifacts.semanticArtifacts.processedAst.cost
+    val alternativeCosts = executionArtifacts.semanticArtifacts.processedAst.alternativeCosts
+
+    return EvalWithCostResult(value, cost, alternativeCosts)
+}
 
 fun evalWithCost(
     architecture: Architecture,
@@ -21,6 +38,29 @@ fun evalWithCost(
 
     val executionScope = ValueTable(globalScope)
 
+    val value = executionArtifacts.processedAst.accept(evalVisitor, EvalContext(executionScope, mapOf()))
+    val cost = executionArtifacts.semanticArtifacts.processedAst.cost
+    val alternativeCosts = executionArtifacts.semanticArtifacts.processedAst.alternativeCosts
+
+    return EvalWithCostResult(value, cost, alternativeCosts)
+}
+
+fun evalWithCost(
+    source: String,
+    architecture: Architecture,
+    sourceStore: SourceStore,
+    pluginSource: String,
+    userPlugins: List<UserPlugin>
+): EvalWithCostResult {
+    val frontend = CompilerFrontend(architecture, sourceStore, UserPluginSource(pluginSource))
+
+    val executionArtifacts = frontend.fullCompileWithTopologicalSort(source)
+
+    val userPluginMap = pluginMap(userPlugins)
+    val globalScope = ValueTable(NullValueTable)
+    val evalVisitor = EvalAstVisitor(architecture, globalScope, userPluginMap)
+
+    val executionScope = ValueTable(globalScope)
     val value = executionArtifacts.processedAst.accept(evalVisitor, EvalContext(executionScope, mapOf()))
     val cost = executionArtifacts.semanticArtifacts.processedAst.cost
     val alternativeCosts = executionArtifacts.semanticArtifacts.processedAst.alternativeCosts
